Cleanup for Native to merge

+++ common/src/utils.rs
@@ -1,10 +1,10 @@
 use futures::Stream;
-use serde::{Deserialize, Serialize};
 use gadget_io::tokio::sync::mpsc::UnboundedReceiver;
+use serde::{Deserialize, Serialize};
 use sp_core::ecdsa;
+use sp_io::EcdsaVerifyError;
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
-use sp_io::EcdsaVerifyError;
 
 pub const ECDSA_SIGNATURE_LENGTH: usize = 65;
 
@@ -96,9 +96,9 @@ pub fn recover_ecdsa_pub_key(
 
         let msg = hash;
         let rid = libsecp256k1::RecoveryId::parse(
-            if sig[64] > 26 { sig[64] - 27 } else { sig[64] } as u8,
+            if sig[64] > 26 { sig[64] - 27 } else { sig[64] },
         )
-            .map_err(|_| EcdsaVerifyError::BadV)?;
+        .map_err(|_| EcdsaVerifyError::BadV)?;
         let sig = libsecp256k1::Signature::parse_overflowing_slice(&sig[..64])
             .map_err(|_| EcdsaVerifyError::BadRS)?;
         let msg = libsecp256k1::Message::parse(&msg);
